+++ CHANGELOG.md
@@ -3,8 +3,8 @@ Changelog
 
 * 1.0.0-alpha2 (@release_date@)
 
- * added `CompositeCommand`
- * added `CompositeCommandApplication`
+ * moved code from `Puli\Cli\Console` to `Puli\Cli`
+ * removed `PuliCli` class
 
 * 1.0.0-alpha1 (2014-12-03)